rename NewGanttChart component
