+++ frontend/src/components/gantt/GanttChart.vue
